% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndr_baixar_precedentes.R
\name{bndr_baixar_precedentes}
\alias{bndr_baixar_precedentes}
\title{Baixar precedentes por id}
\usage{
bndr_baixar_precedentes(id, diretorio = ".")
}
\arguments{
\item{id}{Id extraído por meio da função bnrd_extrair_precedentes}

\item{diretorio}{Onde armazenar os dados}
}
\value{
json
}
\description{
Baixar precedentes por id
}
