% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndr_extrair_precedentes.R
\name{bndr_extrair_precedentes}
\alias{bndr_extrair_precedentes}
\title{Extrai tabela de precedentes}
\usage{
bndr_extrair_precedentes(
  numero = "",
  tipo = "",
  origem_id = "",
  situacao_id = ""
)
}
\arguments{
\item{numero}{Número do precedente. Opcional.}

\item{tipo}{Tipo de precedente: IRDR, RR, IAC, GR, CT, RG
Consulte o dataset tipo para conhecer a descrição.}

\item{origem_id}{Origem do precedente}

\item{situacao_id}{Situação do precedente.}
}
\value{
tibble
}
\description{
Extrai tabela de precedentes
}
