% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndr_extrair_sobrestados.R
\name{bndr_extrair_sobrestados}
\alias{bndr_extrair_sobrestados}
\title{Extrair processos sobrestados}
\usage{
bndr_extrair_sobrestados(id)
}
\arguments{
\item{id}{Id do incidente}
}
\value{
tibble
}
\description{
Extrair processos sobrestados
}
