% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bndr_ler_precedentes.R
\name{bndr_ler_precedentes}
\alias{bndr_ler_precedentes}
\title{Lê dados sobre precedentes baixados com bndr_baixar_precedentes}
\usage{
bndr_ler_precedentes(arquivos = NULL, diretorio = ".")
}
\arguments{
\item{arquivos}{Vetor de arquivos}

\item{diretorio}{Informar diretório se arquivos não forem informados}
}
\value{
tibble
}
\description{
Lê dados sobre precedentes baixados com bndr_baixar_precedentes
}
