% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/origem.R
\docType{data}
\name{origem}
\alias{origem}
\title{origem}
\format{
dataframe com 91 obs e três colunas:
\describe{
\item{\code{id}}{integer id da origem}
\item{\code{descricao}}{character descrição da origem}
\item{\code{codigo_res_65}}{código da origem conforme resolução 65}
}
}
\source{
\url{https://github.com/jjesusfilho/bndr}
}
\usage{
origem
}
\description{
Origem dos incidentes
}
\keyword{datasets}
