% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/situacao.R
\docType{data}
\name{situacao}
\alias{situacao}
\title{situacao}
\format{
dataframe com 77 obs e quatro colunas:
\describe{
\item{\code{tipo}}{character tipo (sigla) de incidente}
\item{\code{id}}{integer id da situação do incidente}
\item{\code{sigla}}{character sigla da situação do incidente}
\item{\code{descricao}}{descrição da situação do incidente}
}
}
\source{
\url{https://github.com/jjesusfilho/bndr}
}
\usage{
situacao
}
\description{
situação dos incidentes conforme os tipos
}
\keyword{datasets}
