% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipo.R
\docType{data}
\name{tipo}
\alias{tipo}
\title{tipo}
\format{
dataframe com 6 obs e duas colunas:
\describe{
\item{\code{sigla}}{character tipo de incidetnte}
\item{\code{descricao}}{descrição do tipo de incidente}
}
}
\source{
\url{https://github.com/jjesusfilho/bndr}
}
\usage{
tipo
}
\description{
tipos de incidentes
}
\keyword{datasets}
